// src/enums/FloorPlanType.ts
// eslint-disable-next-line no-shadow
enum FloorPlanType {
  S1 = 9214480,
  A1 = 9214481,
  A2 = 9214482,
  A3 = 9214483,
  A3L = 9214484,
  A4 = 9214485,
  A4L = 9214486,
  A5 = 9214487,
  B1 = 9214488,
  B2 = 9214489,
  B3 = 9214490,
  B3L = 9214491,
  B4 = 9214492,
  B4L = 9214493,
}

export default FloorPlanType;